requirements.txt
asgi-redis==1.4.3
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.7
djangorestframework==3.11.0
Flask==1.1.2
Flask-SocketIO==4.3.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack-python==0.5.6
numpy==1.18.5
opencv-python==4.2.0.34
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2020.1
redis==2.10.6
requests==2.23.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
Werkzeug==1.0.1
zope.interface==5.1.0
